1. finish connect cache L1 L2 L3. 2. Add cache parameter into instance. 3. calculate num of cache set and new cache set.

+++ CfgParser.h
@@ -19,7 +19,7 @@ class CfgParser {
 
   public:
 	CfgParser( const char * cfg ) ;
-	int ParseDevice( std::string DeviceType, std::string Tag ) ;
+	uint64_t ParseDevice( std::string DeviceType, std::string Tag ) ;
 
 	~CfgParser() ;
 };
